+++ cryptofeed/feed.py
@@ -4,10 +4,11 @@ Copyright (C) 2017-2020  Bryant Moscon - bmoscon@gmail.com
 Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
-import uuid
 from collections import defaultdict
+from typing import Tuple, Callable, Union, List
 
 from cryptofeed.callback import Callback
+from cryptofeed.connection import AsyncConnection
 from cryptofeed.defines import (ASK, BID, BOOK_DELTA, FUNDING, FUTURES_INDEX, L2_BOOK, L3_BOOK, LIQUIDATIONS,
                                 OPEN_INTEREST, MARKET_INFO, TICKER, TRADES, TRANSACTIONS, VOLUME)
 from cryptofeed.exceptions import BidAskOverlapping, UnsupportedDataFeed
@@ -18,7 +19,7 @@ from cryptofeed.util.book import book_delta, depth
 class Feed:
     id = 'NotImplemented'
 
-    def __init__(self, address, pairs=None, channels=None, config=None, callbacks=None, max_depth=None, book_interval=1000, snapshot_interval=False, checksum_validation=False, cross_check=False, origin=None,
+    def __init__(self, address: Union[list, str], pairs=None, channels=None, config=None, callbacks=None, max_depth=None, book_interval=1000, snapshot_interval=False, checksum_validation=False, cross_check=False, origin=None,
                  key_id=None):
         """
         max_depth: int
@@ -39,8 +40,6 @@ class Feed:
         key_id: str
             API key to query the feed, required when requesting supported coins/pairs.
         """
-        self.hash = str(uuid.uuid4())
-        self.uuid = f"{self.id}-{self.hash}"
         self.config = defaultdict(set)
         self.address = address
         self.book_update_interval = book_interval
@@ -93,6 +92,27 @@ class Feed:
         for key, callback in self.callbacks.items():
             if not isinstance(callback, list):
                 self.callbacks[key] = [callback]
+
+    def connect(self) -> List[Tuple[AsyncConnection, Callable[[None], None], Callable[[str, float], None]]]:
+        """
+        Generic connection method for exchanges. Exchanges that require/support
+        multiple addresses will need to override this method in their specific class
+        unless they use the same subscribe method and message handler for all
+        connections.
+
+        Connect returns a list of tuples. Each tuple contains
+        1. an AsyncConnection object
+        2. the subscribe function pointer associated with this connection
+        3. the message handler for this connection
+        """
+        ret = []
+
+        if isinstance(self.address, str):
+            return [(AsyncConnection(self.address, self.id, ping_interval=10, ping_timeout=None, max_size=2**23, max_queue=None, origin=self.origin), self.subscribe, self.message_handler)]
+
+        for n, addr in enumerate(self.address):
+            ret.append((AsyncConnection(addr, self.id, ping_interval=10, ping_timeout=None, max_size=2**23, max_queue=None, origin=self.origin), self.subscribe, self.message_handler))
+        return ret
 
     @classmethod
     def info(cls, key_id: str = None) -> dict:
@@ -194,15 +214,11 @@ class Feed:
     async def message_handler(self, msg: str, timestamp: float):
         raise NotImplementedError
 
+    async def subscribe(self, connection: AsyncConnection):
+        raise NotImplementedError
+
     async def stop(self):
         for callbacks in self.callbacks.values():
             for callback in callbacks:
                 if hasattr(callback, 'stop'):
                     await callback.stop()
-
-
-class RestFeed(Feed):
-    sleep_time = None
-
-    async def message_handler(self):
-        raise NotImplementedError