+++ CHANGES.md
@@ -26,6 +26,8 @@
   * Bugfix: Subscribing to all BitMEX symbols gives 400 error - message too long
   * Bugfix: Cleanup of code - fixed a few examples and resolved all outstanding flake8 issues
   * Bugfix: Fix Bitfinex pair normalization
+  * Feature: Refactor connection handling. New connection design allows feeds to open multiple connections
+  * Feature: Update BitMax to use the new BitMax Pro API - includes sequence number verification on books
 
 ### 1.6.1 (2020-11-12)
   * Feature: New kwarg for exchange feed - `snapshot_interval` - used to control number of snapshot updates sent to client